refactor: Code list page utils
